pyproject.toml
#========================================================================================================================
# Project Settings
#========================================================================================================================
[project]
name = "replace_bixby"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "HenningGC", email = "henning@gruhl.me"},
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

#========================================================================================================================
# Prod Settings
#========================================================================================================================
dependencies = [
    "boto3>=1.34.152",
    "python-dotenv>=1.0.1",
    "pydantic>=2.8.2",
    "pyyaml>=6.0.1",
]

#========================================================================================================================
# PDM Settings
#========================================================================================================================
[tool.pdm]
distribution = false